+++ CORE/SSH/40_SSH.sh
@@ -27,16 +27,30 @@ fi
 rm -rf $HOME/.ssh
 mkdir -p $HOME/.ssh
 
-if [ ! -f $SAVE_DIR/ssh-fs ]; then
-	dd if=/dev/zero of=$SAVE_DIR/ssh-fs bs=1M count=16
-	mkfs.ext4 $SAVE_DIR/ssh-fs
-	sudo mount $SAVE_DIR/ssh-fs $HOME/.ssh
-	sudo chown user:user $HOME/.ssh
-	sudo chmod 755 $HOME/.ssh
-	sudo umount $HOME/.ssh
+# Identify and define the boot device and persistent partition
+BOOT_DEV="$(df -h | grep cdrom | awk -F " " '{print $1}')"
+if [[ $BOOT_DEV == *"nvme"* ]]; then
+	DISK=${BOOT_DEV%??}
+	EXFAT_PARTITION="p3"
+elif [[ $BOOT_DEV == *"mmcblk"* ]]; then
+    DISK=${BOOT_DEV%??}
+	EXFAT_PARTITION="p3"
+else
+	DISK=${BOOT_DEV%?}
+	EXFAT_PARTITION="3"
 fi
 
-sudo mount $SAVE_DIR/ssh-fs $HOME/.ssh
+if [ ! -f $SAVE_DIR/ssh-fs ]; then
+	if [ "${BOOT_DEV}" != "/dev/shm" ] && [ "${BOOT_DEV}" != "/dev/sr0" ] && [ "${BOOT_DEV}" != "/dev/mapper/ventoy" ]; then
+		dd if=/dev/zero of=$SAVE_DIR/ssh-fs bs=1M count=16
+		mkfs.ext4 $SAVE_DIR/ssh-fs
+		sudo mount $SAVE_DIR/ssh-fs $HOME/.ssh
+		sudo chown user:user $HOME/.ssh
+		sudo chmod 755 $HOME/.ssh
+	fi
+else
+	sudo mount $SAVE_DIR/ssh-fs $HOME/.ssh
+fi
 
 } # END OF MODULE COMMANDS FUNCTION
